+++ oneflow/python/model/maskrcnn/maskrcnn_eval.py
@@ -5,6 +5,7 @@ import os
 import numpy as np
 import argparse
 import oneflow as flow
+import oneflow.core.data.data_pb2 as data_util
 
 from datetime import datetime
 from backbone import Backbone
@@ -26,14 +27,49 @@ parser.add_argument("-load", "--model_load_dir", type=str, default="", required=
 parser.add_argument("-g", "--gpu_num_per_node", type=int, default=1, required=False)
 parser.add_argument("-cp", "--ctrl_port", type=int, default=19765, required=False)
 parser.add_argument("-i", "--iter_num", type=int, default=10, required=False)
+parser.add_argument(
+    "-dataset_dir", "--dataset_dir", type=str, default="/dataset/mscoco_2017", required=False
+)
+parser.add_argument(
+    "-anno", "--annotation_file", type=str, default="instances_val2017.json", required=False
+)
+parser.add_argument("-imgd", "--image_dir", type=str, default="val2017", required=False)
+parser.add_argument("-fake", "--fake_img", default=False, action="store_true", required=False)
 terminal_args = parser.parse_args()
 
 
-@flow.function
-def maskrcnn_box_head_eval_job(
-    images=flow.input_blob_def((2, 3, 416, 608), dtype=flow.float, is_dynamic=True),
-    image_sizes=flow.input_blob_def((2, 2), dtype=flow.int32, is_dynamic=False),
+def make_data_loader(
+    batch_size,
+    batch_cache_size=3,
+    dataset_dir="/dataset/mscoco_2017",
+    annotation_file="annotations/sample_2_instances_val2017.json",
+    image_dir="val2017",
+    random_seed=873898,
+    shuffle=False,
+    group_by_aspect_ratio=True,
 ):
+    coco = flow.data.COCODataset(
+        dataset_dir, annotation_file, image_dir, random_seed, shuffle, group_by_aspect_ratio
+    )
+    data_loader = flow.data.DataLoader(coco, batch_size, batch_cache_size)
+    data_loader.add_blob(
+        "image",
+        data_util.DataSourceCase.kImage,
+        shape=(416, 608, 3),
+        dtype=flow.float,
+        is_dynamic=True,
+    )
+    data_loader.add_blob(
+        "image_size", data_util.DataSourceCase.kImageSize, shape=(2,), dtype=flow.int32
+    )
+    data_loader.add_transform(flow.data.TargetResizeTransform(400, 600))
+    data_loader.add_transform(flow.data.ImageNormalizeByChannel((102.9801, 115.9465, 122.7717)))
+    data_loader.add_transform(flow.data.ImageAlign(32))
+    data_loader.init()
+    return data_loader
+
+
+def maskrcnn_box_head_eval(images, image_sizes):
     cfg = get_default_cfgs()
     if terminal_args.config_file is not None:
         cfg.merge_from_file(terminal_args.config_file)
@@ -52,7 +88,7 @@ def maskrcnn_box_head_eval_job(
     for i in range(cfg.DECODER.FPN_LAYERS):
         anchors.append(
             flow.detection.anchor_generate(
-                images=flow.transpose(images, perm=[0, 2, 3, 1]),
+                images=images,
                 feature_map_stride=cfg.DECODER.FEATURE_MAP_STRIDE * pow(2, i),
                 aspect_ratios=cfg.DECODER.ASPECT_RATIOS,
                 anchor_scales=cfg.DECODER.ANCHOR_SCALES * pow(2, i),
@@ -60,7 +96,7 @@ def maskrcnn_box_head_eval_job(
         )
 
     # Backbone
-    features = backbone.build(images)
+    features = backbone.build(flow.transpose(images, perm=[0, 3, 1, 2]))
 
     # RPN
     cls_logit_list, bbox_pred_list = rpn_head.build(features)
@@ -69,7 +105,42 @@ def maskrcnn_box_head_eval_job(
     # Box Head
     cls_probs, box_regressions = box_head.build_eval(proposals, features)
 
-    return tuple(proposals) + tuple(features) + (cls_probs,) + (box_regressions,)
+    return tuple(proposals) + tuple(features) + (image_sizes,) + (cls_probs,) + (box_regressions,)
+
+
+if terminal_args.fake_img:
+
+    @flow.function
+    def maskrcnn_box_head_eval_job(
+        images=flow.input_blob_def(shape=(2, 3, 416, 608), dtype=flow.float32, is_dynamic=True)
+    ):
+        data_loader = make_data_loader(
+            batch_size=terminal_args.batch_size,
+            batch_cache_size=3,
+            dataset_dir=terminal_args.dataset_dir,
+            annotation_file=terminal_args.annotation_file,
+            image_dir=terminal_args.image_dir,
+        )
+        image_sizes = data_loader("image_size")
+
+        return maskrcnn_box_head_eval(flow.transpose(images, perm=[0, 2, 3, 1]), image_sizes)
+
+
+else:
+
+    @flow.function
+    def maskrcnn_box_head_eval_job():
+        data_loader = make_data_loader(
+            batch_size=terminal_args.batch_size,
+            batch_cache_size=3,
+            dataset_dir=terminal_args.dataset_dir,
+            annotation_file=terminal_args.annotation_file,
+            image_dir=terminal_args.image_dir,
+        )
+        images = data_loader("image")
+        image_sizes = data_loader("image_size")
+
+        return maskrcnn_box_head_eval(images, image_sizes)
 
 
 @flow.function
@@ -94,19 +165,23 @@ if __name__ == "__main__":
     flow.config.gpu_device_num(terminal_args.gpu_num_per_node)
     flow.config.ctrl_port(terminal_args.ctrl_port)
     flow.config.default_data_type(flow.float)
+    flow.config.cudnn_conv_heuristic_search_algo(True)
+    flow.config.cudnn_conv_use_deterministic_algo_only(False)
 
     assert terminal_args.model_load_dir != ""
     check_point = flow.train.CheckPoint()
     check_point.load(terminal_args.model_load_dir)
 
-    images = np.load("/tmp/shared_with_jxf/maskrcnn_eval_input_data_small/images.npy")
-    image_sizes = np.load("/tmp/shared_with_jxf/maskrcnn_eval_input_data_small/image_sizes.npy")
-    image_num = image_sizes.shape[0]
-
     # Box Head and Post-Processor
-    results = maskrcnn_box_head_eval_job(images, image_sizes).get()
+    if terminal_args.fake_img:
+        images = np.load("/tmp/shared_with_jxf/maskrcnn_eval_input_data_small/images.npy")
+        results = maskrcnn_box_head_eval_job(images).get()
+    else:
+        results = maskrcnn_box_head_eval_job().get()
+    image_sizes = results[-3]
     cls_probs = results[-2]
     box_regressions = results[-1]
+    image_num = image_sizes.shape[0]
     fpn_feature_map = []
     for i in range(4):
         fpn_feature_map.append(results[image_num + i].ndarray())
@@ -144,6 +219,7 @@ if __name__ == "__main__":
         expected_results=(),
         expected_results_sigma_tol=4,
     )
+
 
 #  Object Detection
 #  Average Precision  (AP) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.474
